requirements.txt
easyocr==1.7.2
numpy==2.2.3
Pillow==11.1.0
